WK-215 moves the presentation components of the DetailPanel in a components folder
